[UI] Исправлена установка токенов на серверной части NextJS Resolve #840
