+++ src/frontend/app/auth/oauth/[type]/oauth.ts
@@ -1,6 +1,6 @@
 "use server";
 import { setTokens } from "@/lib/auth";
-import { OAuth } from "@/lib/client-requests/auth";
+import { OAuth } from "@/lib/server-requests/auth";
 
 export async function login(oauth_type: string, token: string) {
   const tokens = await OAuth(oauth_type.toUpperCase(), token);
